% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regconfint.R
\name{bc_ci}
\alias{bc_ci}
\title{Bias-corrected bootstrap confidence interval}
\usage{
bc_ci(boot.obj, conf = 0.95)
}
\arguments{
\item{boot.obj}{Boot object returned by \code{regconfint()}.}

\item{conf}{Confidence level (default = 0.95).}
}
\value{
Data frame of bias-corrected CIs.
}
\description{
Computes bias-corrected bootstrap confidence intervals from a boot object.
}
