% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regconfint.R
\name{regconfint}
\alias{regconfint}
\title{Estimating confidence interval for regularized regression}
\usage{
regconfint(dataset, expv, tarv, itr, al, sed, fam, lin, typ = "all")
}
\arguments{
\item{dataset}{dataset for calculating CI}

\item{expv}{explanatory variables(character)}

\item{tarv}{target variable(character)}

\item{itr}{number of iteration}

\item{al}{alpha for regularized regression:0(Ridge),1(Lasso) or sequence of alpha for elastic-net}

\item{sed}{seed for bootstrap}

\item{fam}{distribution of the model}

\item{lin}{link function of the model}

\item{typ}{type of intervals required.}
}
\value{
return the bootstrap confidence interval from glmnet object
}
\description{
\code{regconfint} Estimating confidence interval for regularized regression
}
\examples{
#regconfint(dataset=xxxx, expv=c("a","b","c"), tarv="target", itr=1000, al=1, seed=1234, fam="poisson", lin="log")
}
