% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regconfint.R
\name{regconfint}
\alias{regconfint}
\title{Estimate bootstrap confidence interval for regularized regression}
\usage{
regconfint(
  dataset,
  expv,
  tarv,
  itr,
  al,
  sed,
  fam,
  lin,
  typ = "all",
  paral = "N"
)
}
\arguments{
\item{dataset}{data.frame for estimation of confidence intervals}

\item{expv}{vector of explanatory variables}

\item{tarv}{target variable}

\item{itr}{number of iteration}

\item{al}{alpha for regularized regression:0(Ridge),1(Lasso) or sequence of alpha for elastic-net}

\item{sed}{seed for bootstrap}

\item{fam}{Either a character string representing one of the built-in families, or else a glm() family object. For more information, see R documentation of glmnet.}

\item{lin}{link function of the model}

\item{typ}{type of intervals required. Default is "all". For more information, see the R documentation of boot.}

\item{paral}{Y if parallel computing is used for Elastic-net with sequence of alpha. Default is "N"}
}
\value{
return the bootstrap confidence interval from glmnet object
}
\description{
\code{regconfint} regconfint can estimate bootstrap confidence interval for regularized regression
}
\examples{
#Ridge regression
#regconfint(dataset, expv=c("x1","x2"), tarv="y", itr=50, al=0, sed=1, fam=c("gaussian",
#"binomial", "poisson", "multinomial", "cox", "mgaussian")poisson", lin="log")

#Lasso regression
#regconfint(dataset, expv=c("x1","x2"), tarv="y", itr=50, al=1, sed=1, fam="poisson", lin="log")

#Elastic-net with fixed alpha
#regconfint(dataset, expv=c("x1","x2"), tarv="y", itr=50, al=0.5, sed=1, fam="poisson", lin="log")

#Elastic-net with sequence of alpha
#regconfint(dataset, expv=c("x1","x2"), tarv="y", itr=50, al=seq(0.01,0.99,0.01), sed=1, fam="poisson",
# lin="log", paral="N")
# If paral = "Y", parallel computing will be implemented based on the number of cores available
# in your computer
}
